Fix 64-bit WIN32 portability issues, fix msys2/mingw64 build

+++ oniguruma/regint.h
@@ -688,7 +688,7 @@ typedef struct {
   BBuf*  mbuf;   /* multi-byte info or NULL */
 } CClassNode;
 
-typedef long OnigStackIndex;
+typedef size_t OnigStackIndex;
 
 typedef struct _OnigStackType {
   unsigned int type;
@@ -780,7 +780,7 @@ extern int  onig_is_code_in_cc_len P_((int enclen, OnigCodePoint code, CClassNod
 
 /* strend hash */
 typedef void hash_table_type;
-typedef unsigned long hash_data_type;
+typedef size_t hash_data_type;
 
 extern hash_table_type* onig_st_init_strend_table_with_size P_((int size));
 extern int onig_st_lookup_strend P_((hash_table_type* table, const UChar* str_key, const UChar* end_key, hash_data_type *value));
